Renamed the HamburgerMenu to TabMenu
